+++ codemods/misc/migrate-mui-accordion-to-bui-accordion/tests/controlled-todo/expected.tsx
@@ -4,8 +4,8 @@ import AccordionDetails from '@material-ui/core/AccordionDetails';
 
 const MyComponent = ({ expanded, onChange }: any) => (
   <>
-{/* TODO(backstage-codemod): finish accordion migration manually (expanded, onChange) */}
-<Accordion expanded={expanded} onChange={onChange}>
+  {/* TODO(backstage-codemod): finish accordion migration manually (expanded, onChange) */}
+  <Accordion expanded={expanded} onChange={onChange}>
     <AccordionSummary>Settings</AccordionSummary>
     <AccordionDetails>Content</AccordionDetails>
   </Accordion>