+++ codemods/misc/migrate-mui-radio-checkbox-to-bui/tests/complex-form-group-todo/expected.tsx
@@ -4,8 +4,8 @@ import Checkbox from '@material-ui/core/Checkbox';
 
 const MyComponent = () => (
   <>
-{/* TODO(backstage-codemod): finish choice-group migration manually */}
-<FormGroup>
+  {/* TODO(backstage-codemod): finish choice-group migration manually */}
+  <FormGroup>
     <FormControlLabel control={<Checkbox checked={a} onChange={toggleA} />} label="A" />
     <span>Custom separator</span>
   </FormGroup>