+++ codemods/misc/migrate-mui-radio-checkbox-to-bui/tests/partial-barrel-import/input.tsx
@@ -1,0 +1,11 @@
+import { RadioGroup, Radio, FormControlLabel, Button } from '@material-ui/core';
+
+const MyComponent = () => (
+  <>
+    <RadioGroup value={val} onChange={setVal}>
+      <FormControlLabel value="x" control={<Radio />} label="X" />
+      <FormControlLabel value="y" control={<Radio />} label="Y" />
+    </RadioGroup>
+    <Button>Save</Button>
+  </>
+);