+++ codemods/misc/migrate-mui-select-family-to-bui-select/tests/helper-text-todo/expected.tsx
@@ -6,8 +6,8 @@ import FormHelperText from '@material-ui/core/FormHelperText';
 
 const MyComponent = () => (
   <>
-{/* TODO(backstage-codemod): finish Select migration manually */}
-<FormControl>
+  {/* TODO(backstage-codemod): finish Select migration manually */}
+  <FormControl>
     <InputLabel>Color</InputLabel>
     <Select value={color} onChange={e => setColor(e.target.value as string)}>
       <MenuItem value="red">Red</MenuItem>