+++ codemods/misc/migrate-mui-select-family-to-bui-select/tests/multiple-select-todo/expected.tsx
@@ -5,8 +5,8 @@ import MenuItem from '@material-ui/core/MenuItem';
 
 const MyComponent = () => (
   <>
-{/* TODO(backstage-codemod): finish Select migration manually */}
-<FormControl>
+  {/* TODO(backstage-codemod): finish Select migration manually */}
+  <FormControl>
     <InputLabel>Tags</InputLabel>
     <Select multiple value={tags} onChange={handleChange}>
       <MenuItem value="a">A</MenuItem>