+++ codemods/misc/migrate-mui-select-family-to-bui-select/tests/partial-barrel-import/input.tsx
@@ -1,0 +1,14 @@
+import { FormControl, InputLabel, Select, MenuItem, Button } from '@material-ui/core';
+
+const MyComponent = () => (
+  <>
+    <FormControl>
+      <InputLabel>Size</InputLabel>
+      <Select value={size} onChange={e => setSize(e.target.value as string)}>
+        <MenuItem value="sm">Small</MenuItem>
+        <MenuItem value="lg">Large</MenuItem>
+      </Select>
+    </FormControl>
+    <Button>Save</Button>
+  </>
+);