+++ codemods/misc/migrate-mui-slider-to-bui-slider/tests/complex-on-change-todo/expected.tsx
@@ -2,7 +2,7 @@ import Slider from '@material-ui/core/Slider';
 
 const MyComponent = () => (
   <>
-{/* TODO(backstage-codemod): finish slider migration manually (complex-onChange) */}
-<Slider min={0} max={100} onChange={handleSliderChange} />
+  {/* TODO(backstage-codemod): finish slider migration manually (complex-onChange) */}
+  <Slider min={0} max={100} onChange={handleSliderChange} />
 </>
 );