+++ codemods/misc/migrate-mui-slider-to-bui-slider/tests/marks-todo/expected.tsx
@@ -2,7 +2,7 @@ import Slider from '@material-ui/core/Slider';
 
 const MyComponent = () => (
   <>
-{/* TODO(backstage-codemod): finish slider migration manually (marks) */}
-<Slider min={0} max={100} marks step={10} />
+  {/* TODO(backstage-codemod): finish slider migration manually (marks) */}
+  <Slider min={0} max={100} marks step={10} />
 </>
 );