+++ codemods/misc/migrate-mui-slider-to-bui-slider/tests/partial-barrel-import/expected.tsx
@@ -1,0 +1,9 @@
+import { Typography } from '@material-ui/core';
+import { Slider } from '@backstage/ui';
+
+const MyComponent = () => (
+  <>
+    <Typography>Volume</Typography>
+    <Slider minValue={10} maxValue={90} value={50} />
+  </>
+);