+++ codemods/misc/migrate-mui-slider-to-bui-slider/tests/partial-barrel-import/input.tsx
@@ -1,0 +1,8 @@
+import { Slider, Typography } from '@material-ui/core';
+
+const MyComponent = () => (
+  <>
+    <Typography>Volume</Typography>
+    <Slider min={10} max={90} value={50} />
+  </>
+);