+++ codemods/misc/migrate-mui-slider-to-bui-slider/tests/value-label-todo/expected.tsx
@@ -2,7 +2,7 @@ import Slider from '@material-ui/core/Slider';
 
 const MyComponent = () => (
   <>
-{/* TODO(backstage-codemod): finish slider migration manually (valueLabelDisplay) */}
-<Slider min={0} max={100} valueLabelDisplay="auto" />
+  {/* TODO(backstage-codemod): finish slider migration manually (valueLabelDisplay) */}
+  <Slider min={0} max={100} valueLabelDisplay="auto" />
 </>
 );