+++ codemods/misc/migrate-mui-textfield-to-bui-textfield/tests/aliased-import/input.tsx
@@ -1,0 +1,5 @@
+import { TextField as MuiTextField } from '@material-ui/core';
+
+const MyComponent = () => (
+  <MuiTextField label="Title" value={value} onChange={e => setValue(e.target.value)} />
+);