+++ codemods/misc/migrate-mui-textfield-to-bui-textfield/tests/complex-on-change-todo/expected.tsx
@@ -2,7 +2,7 @@ import TextField from '@material-ui/core/TextField';
 
 const MyComponent = () => (
   <>
-{/* TODO(backstage-codemod): finish TextField migration manually (complex-onChange) */}
-<TextField label="Name" value={name} onChange={handleNameChange} />
+  {/* TODO(backstage-codemod): finish TextField migration manually (complex-onChange) */}
+  <TextField label="Name" value={name} onChange={handleNameChange} />
 </>
 );