+++ codemods/misc/migrate-mui-textfield-to-bui-textfield/tests/disabled-and-fullwidth/expected.tsx
@@ -2,7 +2,7 @@ import { TextField } from '@backstage/ui';
 
 const MyComponent = () => (
   <>
-{/* TODO(backstage-codemod): finish TextField migration manually (fullWidth) */}
-<TextField isDisabled label="Status" placeholder="N/A" value={status} />
+  {/* TODO(backstage-codemod): finish TextField migration manually (fullWidth) */}
+  <TextField isDisabled label="Status" placeholder="N/A" value={status} />
 </>
 );