+++ codemods/misc/migrate-mui-textfield-to-bui-textfield/tests/partial-barrel-import/expected.tsx
@@ -1,0 +1,9 @@
+import { Button } from '@material-ui/core';
+import { TextField } from '@backstage/ui';
+
+const MyComponent = () => (
+  <>
+    <TextField label="Title" value={value} onChange={newValue => setValue(newValue)} />
+    <Button>Save</Button>
+  </>
+);